WIP [REACTOS] Fix warning C4146: unary minus operator applied to unsigned type, result still unsigned
Also make it a critical warning

+++ dll/win32/avifil32/api.c
@@ -1765,7 +1765,7 @@ HRESULT WINAPI AVISaveVW(LPCWSTR szFile, CLSID *pclsidHandler,
     goto error;
 
   if (dwInterleave) {
-    LONG lCurFrame = -dwFileInitialFrames;
+    LONG lCurFrame = -(LONG)dwFileInitialFrames;
 
     /* interleaved file */
     if (dwInterleave == 1)
