+++ sdk/lib/drivers/sound/shared/time.c
@@ -29,7 +29,7 @@ SleepMs(ULONG Milliseconds)
 {
     LARGE_INTEGER Period;
 
-    Period.QuadPart = -Milliseconds;
+    Period.QuadPart = -(LONG)Milliseconds;
     Period.QuadPart *= 10000;
 
     KeDelayExecutionThread(KernelMode, FALSE, &Period);
